app searches according to the entered searchterm.
40 results shown instead of 10

Layout formatting, paddings, text sizes, etc.

rotation: The app should save state and restore the list back to the previously scrolled position.

response validation: The app checks whether the device is connected to the internet and responds appropriately. The result of the request is validated to account for a bad server response or lack of server response.

no data: When there is no data to display, the app shows a default TextView that informs the user how to populate the list.